services/ai-service/requirements.txt
pika==1.3.2
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation==0.42b0
scikit-learn==1.3.2
numpy==1.26.2
joblib==1.3.2
requests==2.31.0
